Revert "return single recurring instance"
This reverts commit 60353fd63b9f5111683dfbc7694f446864c67a22.
