+++ src/helpers/event/createEventHelpers/index.ts
@@ -1,2 +1,2 @@
 export { createSingleEvent } from "./createSingleEvent";
-export { createRecurringEvent } from "./createRecurringEvent";
+export { createRecurringEvents } from "./createRecurringEvents";
